Fix various cloud bugs and improve startup
We now wait some time after the machine has started to give it time for
the formatting of the temporary disk. Also, the disk_name variable has
been renamed to better signal which disk it corresponds to.
Finally, the startup script now will be much more careful in the
selection of the disk to format and mount.